requirements.txt
flask
flask-cors
pymongo
bcrypt
jwt
python-dotenv
requests
openai